-> client + server application
Two projects that work togather.
One is a server that listens and the other one is a Client.

The server:
-> reads that from a database
-> receives a list of names from the client
-> checks if there are students who match a given name
-> returns back to the client a boolean array
-> true: there is a student with a name from the list
-> false: no students match a current name

The Client:
-> connects to the server
-> gives the server a list of strings
-> the list of strings is read from a txt file
-> receives the boolean array back with true/false
-> prints the array

Both projects use Sockets to communicate with eachother. The server creates a SocketServer and the client connects to the port the socketServer was created. The database and the jars I took from an older project.